requirements.txt
tensorboard-logger
tensorflow
networkx==3.1
pyemd
matplotlib
scipy
numpy==1.24.0
python-louvain
torchvision
torch_geometric
scikit-learn
